refactor: Move OutputMetric to metrics module.
