+++ pywr-schema/src/lib.rs
@@ -7,6 +7,7 @@
 pub mod data_tables;
 pub mod edge;
 mod error;
+pub mod metric;
 pub mod metric_sets;
 pub mod model;
 pub mod nodes;
